requirements.txt
assemblyai==0.30.0
beautifulsoup4==4.9.1
cryptography==2.9.2
Django>=3.2,<4.0
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.33
langchain_openai==0.1.22
langchain_postgres==0.0.9
openai==1.41.1
python-decouple==3.8
python-dotenv==1.0.1
psycopg2-binary==2.9.6
djangorestframework>=3.12,<3.14
Pillow>=8.2.0,<9.0.0
gunicorn==23.0.0